Separate truth vertex smearing from DST global vertex objects
