+++ simulation/g4simulation/g4vertex/Makefile.am
@@ -6,7 +6,6 @@ AM_CPPFLAGS = \
   -I`root-config --incdir`
 
 lib_LTLIBRARIES = \
-   libg4vertex_io.la \
    libg4vertex.la
 
 AM_LDFLAGS = \
@@ -15,47 +14,22 @@ AM_LDFLAGS = \
 
 libg4vertex_io_la_LIBADD = \
   -lphool \
-  -ltrackbase_historic_io
+  -ltrackbase_historic_io \
+  -lglobalvertex_io
 
 libg4vertex_la_LIBADD = \
-  libg4vertex_io.la \
   -lbbc_io \
   -lg4detectors \
   -lfun4all \
   -lg4bbc_io \
-  -ltrackbase_historic_io
+  -ltrackbase_historic_io \
+  -lglobalvertex_io
 
 pkginclude_HEADERS = \
-  GlobalVertex.h \
-  GlobalVertexv1.h \
-  GlobalVertexFastSimReco.h \
-  GlobalVertexMap.h \
-  GlobalVertexMapv1.h \
-  GlobalVertexReco.h
-
-ROOTDICTS = \
-  GlobalVertex_Dict.cc \
-  GlobalVertexv1_Dict.cc \
-  GlobalVertexMap_Dict.cc \
-  GlobalVertexMapv1_Dict.cc
-
-pcmdir = $(libdir)
-nobase_dist_pcm_DATA = \
-  GlobalVertex_Dict_rdict.pcm \
-  GlobalVertexv1_Dict_rdict.pcm \
-  GlobalVertexMap_Dict_rdict.pcm \
-  GlobalVertexMapv1_Dict_rdict.pcm
-
-libg4vertex_io_la_SOURCES = \
-  $(ROOTDICTS) \
-  GlobalVertex.cc \
-  GlobalVertexv1.cc \
-  GlobalVertexMap.cc \
-  GlobalVertexMapv1.cc
+  GlobalVertexFastSimReco.h
 
 libg4vertex_la_SOURCES = \
-  GlobalVertexFastSimReco.cc \
-  GlobalVertexReco.cc
+  GlobalVertexFastSimReco.cc
 
 # Rule for generating table CINT dictionaries.
 %_Dict.cc: %.h %LinkDef.h
@@ -75,7 +49,6 @@ BUILT_SOURCES = \
   testexternals.cc
 
 testexternals_g4vertex_io_SOURCES = testexternals.cc
-testexternals_g4vertex_io_LDADD = libg4vertex_io.la
 
 testexternals_g4vertex_SOURCES = testexternals.cc
 testexternals_g4vertex_LDADD = libg4vertex.la
